iOS: Suppress warnings about unused functions
There are a fair number of warnings about unused functions, mostly in
third-party modules, and we would like the build to be warning free.

We need to change the main Makefile, as -Wall is currently added in
Makerules, and we need our -Wno-unused-function in XCFLAGS to be
after -Wall on the command line for it to have any effect.